feat(6907): extract m4 escrow integration modules
